[libffi] Fix build when cross-compilation
Trying to cross-compile libffi on Windows x86_64 for e.g. arm64
currently fails, because CMake is not aware of a cross-compilation build
and thus CMAKE_SYSTEM_PROCESSOR is set to the host processor, i.e.
AMD64. This causes the wrong TARGET to be selected, which in turn causes
the x86 assembler ml to be selected. However, this assembler is not
in the current search path causing the build to fail, complaining that
sysv_intel.obj cannot be found.

To fix this it seems that the usage of VCPKG_TARGET_ARCHITECTURE is
appropriate.